+++ srcjs/shinydashboardPlus-2.0.0/controlbar.js
@@ -37,22 +37,19 @@ $(function() {
 
   Shiny.inputBindings.register(controlbarBinding, "controlbar-input");
 
-  var controlbarChange = function() {
-    // 1) Trigger the resize event (so images are responsive and resize)
-    $(window).trigger("resize");
-  };
-
-  // Whenever the right sidebar (controlbar) finishes a transition (which it does every time it
-  // changes from collapsed to expanded and vice versa), call controlbarChange()
-  $(".control-sidebar").on(
-    "webkitTransitionEnd otransitionend oTransitionEnd msTransitionEnd transitionend",
-    controlbarChange
-  );
+  // This prevent box content from going outside their container 
+  // when the control-bar is on push mode
+  $("[data-toggle='control-sidebar']").on("click",
+    function() {
+      if (!$.AdminLTE.options.controlSidebarOptions.slide) {
+        $(window).trigger("resize"); 
+      }
+  });
 
   // this step is to overwrite global adminLTE options
   // to set the controlbar slide value
-  var overlay = $(".control-sidebar").attr("data-overlay") === "true";
-  $.AdminLTE.options.controlSidebarOptions.slide = overlay;
+  //var overlay = $(".control-sidebar").attr("data-overlay") === "true";
+  //$.AdminLTE.options.controlSidebarOptions.slide = overlay;
 
   // toggle controlbar at start
   var controlbarCollapsed = $(".control-sidebar").attr("data-collapsed");