+++ srcjs/shinydashboardPlus-2.0.0/miscellaneous.js
@@ -1,4 +1,5 @@
 $(function() {
+  
   // box vertical overflow
   $(".content-wrapper").css("overflow-y", "auto");
 
@@ -60,16 +61,6 @@ $(function() {
   $(".carousel").on("slide.bs.carousel", function() {
     $(this).trigger("shown");
   });
-
-  // overwrite box animation speed. Putting 500 ms add unnecessary delay for Shiny.
-  $.AdminLTE.boxWidget.animationSpeed = 10; 
-  hasGlobalConfig = $(document).find("script[data-for='adminLTEConfig']");
-  if (hasGlobalConfig.length > 0) {
-    globalConfig = JSON.parse(hasGlobalConfig.html());
-    if (globalConfig.boxWidget.animationSpeed !== undefined) {
-      $.AdminLTE.boxWidget.animationSpeed = globalConfig.boxWidget.animationSpeed;
-    }
-  }
 
   /**
    * List of all the available skins